Support energy expressions to be spintraced
The closed_shell_CC_spintrace function would error if it was fed the
CC energy equations, because it unconditionally performs a biorthogonal
transform, which assumes the presence of external indices (aka: it
assumes that we are dealing with a CC residual equation).

This commit changes the code so that we only perform the transformation
for residuals, but skip it for energy expressions (fully contracted
expressions).