Revamp and fix enable-ifs in expr.cpp
The used enable-ifs looked a bit scary and have therefore been
refactored to use intermediates with (hopefully) clear names.

Furthermore, a couple of tests in those enable-ifs have been performed
with std::is_same, but only removing const from the compared type and
thus these checks would not work properly as soon as a reference type
was passed.
This could be observed while constructing a Product from a non-const
other product. Thus, the type of the to-be-copied object was just
Product & and not const Product &, which is what the copy-ctor
(correctly) is defined for. However, there is the templated
Product(Range &&) ctor, that thanks to the universal reference is the
more specialized candidate in this case (it takes the type as-is without
requiring a const to be slapped on). This would lead to the product
being copied, but also flattened (as this is the default behavior of the
range-ctor).